+++ protocols/index.js
@@ -39,6 +39,7 @@ import quake2 from './quake2.js'
 import quake3 from './quake3.js'
 import rfactor from './rfactor.js'
 import samp from './samp.js'
+import satisfactory from './satisfactory.js'
 import savage2 from './savage2.js'
 import starmade from './starmade.js'
 import starsiege from './starsiege.js'
@@ -67,7 +68,7 @@ export {
   armagetron, ase, asa, assettocorsa, battlefield, buildandshoot, cs2d, discord, doom3, eco, epic, factorio, farmingsimulator, ffow,
   fivem, gamespy1, gamespy2, gamespy3, geneshift, goldsrc, gtasao, hexen2, jc2mp, kspdmp, mafia2mp, mafia2online, minecraft,
   minecraftbedrock, minecraftvanilla, minetest, mumble, mumbleping, nadeo, openttd, palworld, quake1, quake2, quake3, rfactor, ragemp, samp,
-  savage2, starmade, starsiege, teamspeak2, teamspeak3, terraria, tribes1, tribes1master, unreal2, ut3, valve,
+  satisfactory, savage2, starmade, starsiege, teamspeak2, teamspeak3, terraria, tribes1, tribes1master, unreal2, ut3, valve,
   vcmp, ventrilo, warsow, eldewrito, beammpmaster, beammp, dayz, theisleevrima, xonotic, altvmp, vintagestorymaster,
   vintagestory
 }