+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 5.X.Y
 * Feat: Replaced `punycode` package usage with `url.domainToASCII` (#630).
 * Feat: World of Padman (2007) - Added support (#636)
+* Feat: Satisfactory - Added support (#442) (By @Smidy13)
 
 ## 5.1.3
 * Fix: `Deus Ex` using the wrong protocol (#621)